Fixed a problem where only normal blocks where spamming power ups and starting preparing the block grid to be dinamically created.

+++ Assets/Scripts/GameController.cs
@@ -19,7 +19,7 @@ public class GameController : MonoBehaviour {
     private int lives;
     public Text livesText;
     public float initBallSpeed;
-    public int poweUpChancePerc;
+    public float poweUpChancePerc;
     public int powerUpSpeed;
 
 	// Awake is called when the script instance is being loaded.
